Moved SlogBackedGinLogger to middleware folder/package
